requirements.txt
typer==0.9.0
python-dotenv==1.0.0
httpx==0.25.1
pydantic==2.5.1